Monsters can be hurt & killed
BUG: a tile is blocked 1 turn after the monster is dead

+++ src/components/combat.rs
@@ -4,7 +4,7 @@ use specs_derive::Component;
 #[derive(Component, Debug)]
 pub struct CombatStats {
     pub max_hp: u16,
-    pub hp: u16,
+    pub hp: i32,
     pub defense: i32,
     pub power: u16,
 }
